refactor(S-19.03): hoist check_path_allowed to path_util (dedup read/write gate; F-P2-001)
