+++ crates/factory-dispatcher/src/host/read_file.rs
@@ -18,7 +18,7 @@ use serde_json::{Map, Value};
 use wasmtime::Linker;
 
 use super::memory::{read_wasm_string, write_wasm_bytes, write_wasm_u32};
-use super::path_util::{PathAllowDecision, resolve_path_for_allowlist};
+use super::path_util::{PathAllowDecision, check_path_allowed};
 use super::{HostCallError, HostCaller, HostContext, codes};
 use crate::internal_log::InternalEvent;
 
@@ -152,54 +152,6 @@ fn resolve_for_read(path: &Path, base: &Path) -> PathBuf {
     } else {
         base.join(path)
     }
-}
-
-/// Two-step allowlist check (architect Ruling-2 / S-19.03 AC-001).
-///
-/// Step 1: resolve via ancestor-walk+rejoin (handles absent files correctly,
-///   unlike `Path::canonicalize()` which returns Err for non-existent files).
-///   Returns `DeniedResolutionFailed` when even the root ancestor fails —
-///   structurally impossible on real Unix filesystems, but testable via the
-///   injectable mock in path_util tests (BC-2.07.001 EC-007).
-///
-/// Step 2: pure `starts_with` prefix check against each allow-list entry.
-///   Allow-list entries that are relative are expanded under `base`.
-///   Returns `DeniedNotAllowed` when the resolved path lies outside all prefixes.
-///
-/// Separating resolution failure from allowlist failure lets operators distinguish
-/// filesystem errors from genuine access-policy violations in telemetry.
-pub(crate) fn check_path_allowed(
-    resolved: &Path,
-    allow: &[String],
-    base: &Path,
-    canonicalize_fn: impl Fn(&Path) -> std::io::Result<PathBuf> + Copy,
-) -> PathAllowDecision {
-    // Step 1: resolve with ancestor-walk+rejoin so absent-but-allowlisted files
-    // get a synthesized canonical path instead of an opaque resolution failure.
-    let canon_resolved = match resolve_path_for_allowlist(resolved, canonicalize_fn) {
-        Some(p) => p,
-        None => return PathAllowDecision::DeniedResolutionFailed,
-    };
-
-    // Step 2: prefix check. Allow-list entries are also resolved via ancestor-walk+rejoin
-    // so that file-scoped allow-list entries (e.g. ".factory/wave-state.yaml") work
-    // correctly even when the file does not yet exist. If the prefix's entire ancestor
-    // chain fails canonicalization, that prefix is skipped.
-    for pref in allow {
-        let pref_path = if Path::new(pref).is_absolute() {
-            PathBuf::from(pref)
-        } else {
-            base.join(pref)
-        };
-        let canon_pref = match resolve_path_for_allowlist(&pref_path, canonicalize_fn) {
-            Some(p) => p,
-            None => continue, // configured prefix's ancestors also absent — skip
-        };
-        if canon_resolved.starts_with(&canon_pref) {
-            return PathAllowDecision::Allowed;
-        }
-    }
-    PathAllowDecision::DeniedNotAllowed
 }
 
 fn read_bounded(path: &Path, max_bytes: usize) -> Result<Vec<u8>, ReadErr> {
